# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "What Deck API",
        "description": "Magic: The Gathering card collection management API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/collections/{id}/import/moxfield": {
            "post": {
                "tags": [
                    "Import"
                ],
                "summary": "Import Moxfield CSV",
                "description": "Import cards from a Moxfield CSV export into a specific collection",
                "operationId": "e9b83a1bc7f6da2d31eb409a1f3909c2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Collection ID to import cards into",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "csv_file"
                                ],
                                "properties": {
                                    "csv_file": {
                                        "description": "Moxfield CSV export file",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "create_missing_cards": {
                                        "description": "Whether to create new card records for unknown cards",
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Import completed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Import completed successfully"
                                        },
                                        "stats": {
                                            "properties": {
                                                "processed": {
                                                    "type": "integer",
                                                    "example": 150
                                                },
                                                "cards_created": {
                                                    "type": "integer",
                                                    "example": 75
                                                },
                                                "cards_found": {
                                                    "type": "integer",
                                                    "example": 75
                                                },
                                                "instances_created": {
                                                    "type": "integer",
                                                    "example": 150
                                                },
                                                "errors": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "row": {
                                                                "type": "integer",
                                                                "example": 5
                                                            },
                                                            "message": {
                                                                "type": "string",
                                                                "example": "Invalid condition value"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid file or import error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Collection not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ValidationError": {
                "title": "Validation Error Response",
                "description": "Response format for validation errors",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "title": [
                                "The title field is required."
                            ],
                            "type": [
                                "The type field is required."
                            ]
                        },
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "title": "Error Response",
                "description": "Standard error response format",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Resource not found"
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "card": [
                                "This card has existing instances and cannot be deleted"
                            ]
                        },
                        "nullable": true,
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}